# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/persons": {
            "get": {
                "description": "Get a list of all persons",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "persons"
                ],
                "summary": "Retrieve all persons",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/domain.Person"
                            }
                        }
                    }
                }
            }
        },
        "/persons/search": {
            "get": {
                "description": "Search for persons using a text query",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "persons"
                ],
                "summary": "Search for persons",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Text to search for",
                        "name": "text",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/domain.Person"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "domain.ContactInfo": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                }
            }
        },
        "domain.Person": {
            "type": "object",
            "properties": {
                "about": {
                    "type": "string"
                },
                "contacts": {
                    "$ref": "#/definitions/domain.ContactInfo"
                },
                "department": {
                    "type": "string"
                },
                "division": {
                    "type": "string"
                },
                "head": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "jobtitle": {
                    "type": "string"
                },
                "middle_name_rus": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "surname": {
                    "type": "string"
                },
                "team": {
                    "type": "string"
                },
                "working_hour": {
                    "type": "string"
                },
                "workplace": {
                    "type": "string"
                }
            }
        }
    }
}